requirements.txt
paho-mqtt
PyQt5
